add development enhancements
